+++ lib/appium_lib/driver.rb
@@ -3,7 +3,25 @@ module Minitest
   # Fix superclass mismatch for class Spec
   class Runnable
   end
-  class Test < Runnable
+
+  # To switch load class
+  def self.left_greater_than_or_equal_to_right?(left, right)
+    left.split('.').zip(right.split('.')).each do |value|
+      diff = value[0].to_i - value[1].to_i
+      return true if diff > 0
+      return false if diff < 0
+    end
+    true
+  end
+
+  if left_greater_than_or_equal_to_right?(VERSION, '5.11.0')
+    # http://docs.seattlerb.org/minitest/History_rdoc.html#label-5.11.0+-2F+2018-01-01
+    # `Minitest::Test` became a subclass of `Minitest::Result`
+    class Test < Result
+    end
+  else
+    class Test < Runnable
+    end
   end
   class Spec < Test
   end